use ErrorTypeError instead
