+++ frontend/packages/@depmap/dataset-manager/src/components/DatasetForm.tsx
@@ -16,7 +16,7 @@ import {
   SampleDimensionType,
   FeatureDimensionType,
   Dataset,
-  instanceOfErrorDetail,
+  ErrorTypeError,
 } from "@depmap/types";
 import ChunkedFileUploader from "./ChunkedFileUploader";
 import { CeleryTask } from "@depmap/compute";
@@ -165,13 +165,13 @@ export default function DatasetForm(props: DatasetFormProps) {
     const isCeleryTask = (x: any): x is CeleryTask => x.state !== undefined;
     if (isCeleryTask(res)) {
       setCompletedTask(res);
-    } else if (instanceOfErrorDetail(res)) {
+    } else if (e instanceof ErrorTypeError) {
       // can occur when error happens before task passed to celery (ex: 'units' missing in 'continuous' col_type)
       setCompletedTask({
         id: "",
         state: "FAILURE",
         percentComplete: undefined,
-        message: typeof res.message,
+        message: res.message,
         result: null,
       });
     } else {
