+++ frontend/packages/@depmap/types/index.ts
@@ -94,5 +94,4 @@ export type { SliceQuery } from "./src/SliceQuery";
 export type { ErrorDetail } from "./src/BreadboxErrorType";
 
 // predicates (these are functions, not types, but they help to narrow types)
-export { instanceOfErrorDetail } from "./src/BreadboxErrorType";
 export { isValidSliceQuery, areSliceQueriesEqual } from "./src/SliceQuery";
